requirements.txt
streamlit
pandas
matplotlib
plotly
openpyxl
fpdf2
google-generativeai
